# perf: optimize cli loading