--- a/internal/core/bootstrap.go
+++ b/internal/core/bootstrap.go
@@ -187,9 +187,9 @@ func Bootstrap(config *BootstrapConfig) (exitCode int, result interface{}, err e
 
 	// cobraBuilder will build a Cobra root command from a list of Command
 	builder := cobraBuilder{
-		commands: config.Commands.GetSortedCommand(),
-		meta:     meta,
-		ctx:      ctx,
+		cmds: config.Commands,
+		meta: meta,
+		ctx:  ctx,
 	}
 
 	rootCmd := builder.build()