Base run
```
 ± time scw -h                
...
scw -h  0.70s user 0.05s system 172% cpu 0.434 total
```

Current
```
 ± time ./scw -h              
...
./scw -h  0.15s user 0.04s system 162% cpu 0.114 total
```