@Codelax 2022-12-26
After https://github.com/scaleway/scaleway-cli/pull/2697/commits/246cc4fbb05ddfbac5bedf3cda6cc1f2c12155d9
```
 ± time ./scw -h
...
./scw -h  0.05s user 0.03s system 119% cpu 0.064 total
```